checkpoint: tag all tests that depend on a running pachd as "livek8s"; add the ability to skip local docker tests
