+++ src/client/godoc_test.go
@@ -1,3 +1,6 @@
+//go:build livek8s
+// +build livek8s
+
 //nolint:wrapcheck
 package client
