+++ src/internal/testutil/enterprise.go
@@ -14,7 +14,8 @@ import (
 func GetTestEnterpriseCode(t testing.TB) string {
 	acode, exists := os.LookupEnv("ENT_ACT_CODE")
 	if !exists {
-		t.Error("Enterprise Activation code not found in Env Vars")
+		t.Log("Enterprise Activation code not found in environment variable ENT_ACT_CODE")
+		t.SkipNow()
 	}
 	return acode
