+++ src/internal/testutil/kubeclient_live.go
@@ -1,0 +1,36 @@
+//go:build livek8s
+// +build livek8s
+
+package testutil
+
+import (
+	"os"
+	"testing"
+
+	"github.com/pachyderm/pachyderm/v2/src/internal/require"
+
+	kube "k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+)
+
+// GetKubeClient connects to the Kubernetes API server either from inside the
+// cluster or from a test binary running on a machine with kubectl (it will
+// connect to the same cluster as kubectl)
+func GetKubeClient(t testing.TB) *kube.Clientset {
+	var config *rest.Config
+	var err error
+	host := os.Getenv("KUBERNETES_SERVICE_HOST")
+	if host != "" {
+		config, err = rest.InClusterConfig()
+	} else {
+		rules := clientcmd.NewDefaultClientConfigLoadingRules()
+		kubeConfig := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(rules,
+			&clientcmd.ConfigOverrides{})
+		config, err = kubeConfig.ClientConfig()
+	}
+	require.NoError(t, err)
+	k, err := kube.NewForConfig(config)
+	require.NoError(t, err)
+	return k
+}
