+++ src/internal/testutil/kubeclient_nonlive.go
@@ -1,0 +1,16 @@
+//go:build !livek8s
+// +build !livek8s
+
+package testutil
+
+import (
+	"testing"
+
+	kube "k8s.io/client-go/kubernetes"
+)
+
+func GetKubeClient(t testing.TB) *kube.Clientset {
+	t.Helper()
+	t.Fatal("a kubernetes client can only be retrived when built with the 'livek8s' build tag")
+	return nil
+}
