+++ src/internal/testutil/pachclient.go
@@ -1,0 +1,3 @@
+package testutil
+
+const DefaultTransformImage = "pachyderm/testuser:local"
