+++ src/server/auth/server/testing/admin_test.go
@@ -1,3 +1,6 @@
+//go:build livek8s
+// +build livek8s
+
 // admin_test.go tests various features related to pachyderm's auth admins.
 // Because the cluster has one global set of admins, these tests can't be run in
 // parallel
