+++ src/server/license/server/license_test.go
@@ -1,3 +1,6 @@
+//go:build livek8s
+// +build livek8s
+
 package server
 
 import (
