+++ src/server/pfs/cmds/mount_linux.go
@@ -1,3 +1,4 @@
+//go:build linux
 // +build linux
 
 package cmds
