+++ src/server/pfs/cmds/mount_unix.go
@@ -1,3 +1,4 @@
+//go:build !windows
 // +build !windows
 
 package cmds
