+++ src/server/pfs/cmds/mount_windows.go
@@ -1,3 +1,4 @@
+//go:build windows
 // +build windows
 
 package cmds
