+++ src/server/pps/server/s3g_sidecar_test.go
@@ -1,3 +1,6 @@
+//go:build livek8s
+// +build livek8s
+
 /*
 This test is for PPS pipelines that use S3 inputs/outputs. Most of these
 pipelines use the pachyderm/s3testing image, which exists on dockerhub but can
